requirements.txt
PyQt5
requests
google-api-python-client
google-auth
google-auth-oauthlib